requirements.txt
pandas
numpy
scikit-learn
joblib
matplotlib
seaborn
flask
flask-cors
google-generativeai
python-dotenv
